+++ cart/views.py
@@ -2,10 +2,13 @@ from django.shortcuts import render
 from django.shortcuts import render, redirect
 from .models import *
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 
 def home(request):
     return redirect('../')
 
+@login_required
 def carrello(request):
     
     templ="cart/carrello.html"
@@ -43,15 +46,19 @@ def carrello(request):
 
 def delete_item(request,prod_code):
 
-    user=User.objects.get(username=request.user) #acquisisco user
-    client=Client.objects.get(user=user) #acquisisco client
-    carrello=Carrello.objects.get(user=client) #Acquisisco carrello relativo all'utente loggato
+    if Product.objects.filter(product_code=prod_code).exists():
+        user=User.objects.get(username=request.user) #acquisisco user
+        client=Client.objects.get(user=user) #acquisisco client
+        carrello=Carrello.objects.get(user=client) #Acquisisco carrello relativo all'utente loggato
 
-    for c in carrello.prodotti.all():
-        if c.item.product_code == prod_code:
-            c.delete() #elimino oggetto dal carrello
+        for c in carrello.prodotti.all():
+            if c.item.product_code == prod_code:
+                c.delete() #elimino oggetto dal carrello
 
-    return redirect("../carrello")
+        return redirect("../carrello")
+
+    else:
+         return HttpResponse("ERROR: product_code non valido")
